requirements.txt
pandas
pyyaml
openpyxl
rich
numpy
